HS-106 As a Recruitment Manager, I want to find candidates which have the same job skills

+++ tn.esprit.R2S-ejb/src/main/java/tn/esprit/R2S/interfaces/ICandidateService.java
@@ -19,7 +19,7 @@ public interface ICandidateService {
 
     List<Candidate> findAll();
 
-    List<Candidate> findBySkillId(int skillId);
+    List<Candidate> findBySkillId(Long skillId);
 
     List<Candidate> findByExperience(int duration);
